requirements.txt
aiogram>=3.13.0
aiohttp>=3.10.5
python-dotenv>=1.0.1
Pillow>=11.0.0
asyncio-mqtt==0.16.2